refactor: Split Balance
