+++ src/ducks/balance/Balance.styl
@@ -1,89 +1,4 @@
 @require 'settings/breakpoints'
-@require '~styles/mixins.styl'
-@require 'settings/palette.styl'
-
-.Balance__table
-    @extend $full-width-table
-
-.Balance__row
-    cursor pointer
-    composes mobile-row from '../../components/Table/styles.styl'
-    composes nav-row from '../commons/mobile.styl'// @stylint ignore
-
-    +small-screen()
-        &
-            padding-right 2.5rem
-
-.Balance__table
-    .Balance__row
-        &:active
-        &:focus
-            background zircon
-
-.Balance__row--selected
-    border-left 0.25rem dodgerBlue solid
-    background zircon
-
-.Balance__row--selected-account-from-group
-    border-left 0.25rem #cdced2 solid
-    background paleGrey
-
-.Balance__account_name
-    maxed-flex-basis 25%
-
-.Balance__solde
-    maxed-flex-basis 20%
-    text-align right
-
-.Balance__bank
-    maxed-flex-basis 30%
-    color coolGrey
-    composes mobile-row-label from '../../components/Table/styles.styl'
-
-
-
-.Balance__group-accounts
-.Balance__account_number
-    maxed-flex-basis 25%
-    color coolGrey
-    composes mobile-row-label from '../../components/Table/styles.styl'
-
-.Balance__table
-    @extend $full-width-table
-    tbody
-        .Balance__solde
-        .Balance__account_name
-            color charcoalGrey
-
-            &.warning
-                font-weight bold
-                color texasRose
-
-                span
-                    margin-left 5px
-                    heigth 1rem
-
-                    &:before
-                        vertical-align middle
-                        margin-bottom 3px
-            &.alert
-                color pomegranate
-
-+small-screen()
-    td.Balance__account_name
-        maxed-flex-basis 60%
-        padding-top 0
-        padding-bottom 0
-
-    td.Balance__solde
-        maxed-flex-basis 40%
-
-    td.Balance__solde, td.Balance__account_name
-        padding-top .5rem
-        padding-bottom .5rem
-
-    .Balance__bank
-        maxed-flex-basis 60%
 
 .Balance_Header_WithChart
     min-height 245px
