+++ src/ducks/balance/components/BalanceAccounts.jsx
@@ -1,0 +1,53 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import cx from 'classnames'
+import { translate, Button } from 'cozy-ui/react'
+
+import AddAccountLink from 'ducks/settings/AddAccountLink'
+import btnStyles from 'styles/buttons.styl'
+
+import BalanceTable from './BalanceTable'
+import BalanceRow from './BalanceRow'
+
+class BalanceAccounts extends React.PureComponent {
+  render() {
+    const { accounts, balanceLower, t } = this.props
+    const names = [
+      t('Accounts.label'),
+      t('Balance.solde'),
+      t('Balance.account_number'),
+      t('Balance.bank_name')
+    ]
+
+    return (
+      <div>
+        <h3>{t('AccountSwitch.accounts')}</h3>
+        <BalanceTable names={names}>
+          {accounts.map((account, index) => (
+            <BalanceRow
+              key={index}
+              account={account}
+              warningLimit={balanceLower}
+            />
+          ))}
+        </BalanceTable>
+        <p>
+          <AddAccountLink>
+            <Button
+              className={cx(btnStyles['btn--no-outline'], 'u-pv-1')}
+              icon="plus"
+              label={t('Accounts.add-account')}
+            />
+          </AddAccountLink>
+        </p>
+      </div>
+    )
+  }
+}
+
+BalanceAccounts.propTypes = {
+  balanceLower: PropTypes.number.isRequired,
+  accounts: PropTypes.array.isRequired
+}
+
+export default translate()(BalanceAccounts)
