+++ src/ducks/balance/components/BalanceRow.styl
@@ -1,0 +1,42 @@
+@require 'settings/breakpoints'
+@require '~styles/mixins.styl'
+@require 'settings/palette.styl'
+
+.BalanceRow
+    &:active
+    &:focus
+        background zircon
+
+    cursor pointer
+    composes mobile-row from '../../../components/Table/styles.styl'
+    composes nav-row from '../../commons/mobile.styl' // @stylint ignore
+
+    +small-screen()
+        padding-right 2.5rem
+
+    td
+        +small-screen()
+            padding-top: 0.5rem;
+            padding-bottom: 0.5rem;
+
+        &.warning
+            font-weight bold
+            color texasRose
+
+            span
+                margin-left 5px
+                heigth 1rem
+
+                &:before
+                    vertical-align middle
+                    margin-bottom 3px
+        &.alert
+            color pomegranate
+
+.BalanceRow--selected
+    border-left 0.25rem dodgerBlue solid
+    background zircon
+
+.BalanceRow--selected-account-from-group
+    border-left 0.25rem #cdced2 solid
+    background paleGrey
