+++ src/ducks/balance/components/BalanceTable.styl
@@ -1,0 +1,38 @@
+@require 'settings/breakpoints'
+@require '~styles/mixins.styl'
+
+.BalanceTable
+    @extend $full-width-table
+
+    tbody
+        .ColumnName
+        .ColumnSolde
+            color charcoalGrey
+
+.ColumnName
+    maxed-flex-basis 25%
+    color charcoalGrey
+
+    +small-screen()
+        maxed-flex-basis 60%
+
+.ColumnSolde
+    maxed-flex-basis 20%
+    text-align right
+
+    +small-screen()
+        maxed-flex-basis 40%
+
+.ColumnAccount
+    maxed-flex-basis 25%
+    color coolGrey
+    composes mobile-row-label from '../../../components/Table/styles.styl'
+
+.ColumnBank
+    maxed-flex-basis 30%
+    color coolGrey
+    composes mobile-row-label from '../../../components/Table/styles.styl'
+
+    +small-screen()
+        .Balance__bank
+            maxed-flex-basis 60%
