+++ src/ducks/balance/components/index.js
@@ -1,0 +1,4 @@
+import BalanceGroups from './BalanceGroups'
+import BalanceAccounts from './BalanceAccounts'
+
+export { BalanceAccounts, BalanceGroups }
